feat(dui3): blocks poc wip: adds rhino layer manager, moves instance baking to instance manager, etc.
