+++ DUI3-DX/Connectors/Rhino/Speckle.Connectors.Rhino7/DependencyInjection/RhinoConnectorModule.cs
@@ -73,5 +73,6 @@ public class RhinoConnectorModule : ISpeckleModule
 
     builder.AddScoped<IRootObjectBuilder<RhinoObject>, RhinoRootObjectBuilder>();
     builder.AddScoped<IInstanceObjectsManager<RhinoObject>, RhinoInstanceObjectsManager>();
+    builder.AddScoped<RhinoLayerManager>();
   }
 }
