+++ DUI3-DX/Connectors/Rhino/Speckle.Connectors.Rhino7/Operations/Receive/RhinoHostObjectBuilder.cs
@@ -1,9 +1,9 @@
-using System.Diagnostics.Contracts;
 using System.DoubleNumerics;
 using Rhino;
 using Rhino.DocObjects;
 using Rhino.Geometry;
 using Speckle.Connectors.Rhino7.Extensions;
+using Speckle.Connectors.Rhino7.HostApp;
 using Speckle.Connectors.Utils.Builders;
 using Speckle.Connectors.Utils.Conversion;
 using Speckle.Connectors.Utils.Instances;
@@ -22,17 +22,20 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
   private readonly IConversionContextStack<RhinoDoc, UnitSystem> _contextStack;
   private readonly GraphTraversal _traverseFunction;
   private readonly IInstanceObjectsManager<RhinoObject> _instanceObjectsManager;
+  private readonly RhinoLayerManager _layerManager;
 
   public RhinoHostObjectBuilder(
     IRootToHostConverter converter,
     IConversionContextStack<RhinoDoc, UnitSystem> contextStack,
     GraphTraversal traverseFunction,
+    RhinoLayerManager layerManager,
     IInstanceObjectsManager<RhinoObject> instanceObjectsManager
   )
   {
     _converter = converter;
     _contextStack = contextStack;
     _traverseFunction = traverseFunction;
+    _layerManager = layerManager;
     _instanceObjectsManager = instanceObjectsManager;
   }
 
@@ -67,13 +70,7 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     return conversionResults;
   }
 
-  // POC: Potentially refactor out into an IObjectBaker.
-  // POC: temp disable
-#pragma warning disable CA1506
-#pragma warning disable CA1502
   private HostObjectBuilderResult BakeObjects(
-#pragma warning restore CA1502
-#pragma warning restore CA1506
     IEnumerable<TraversalContext> objectsGraph,
     List<InstanceDefinitionProxy>? instanceDefinitionProxies,
     string baseLayerName,
@@ -81,13 +78,10 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
   )
   {
     RhinoDoc doc = _contextStack.Current.Document;
+
     var rootLayerIndex = _contextStack.Current.Document.Layers.Find(Guid.Empty, baseLayerName, RhinoMath.UnsetIntIndex);
-
     PreReceiveDeepClean(baseLayerName, rootLayerIndex);
-
-    var cache = new Dictionary<string, int>();
-    rootLayerIndex = doc.Layers.Add(new Layer { Name = baseLayerName });
-    cache.Add(baseLayerName, rootLayerIndex);
+    _layerManager.CreateBaseLayer(baseLayerName);
 
     using var noDraw = new DisableRedrawScope(doc.Views);
 
@@ -108,7 +102,7 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
 
     foreach (TraversalContext tc in traversalContexts)
     {
-      var path = GetLayerPath(tc);
+      var path = _layerManager.GetLayerPath(tc);
       if (tc.Current is IInstanceComponent flocker)
       {
         instanceComponents.Add((path, flocker));
@@ -120,20 +114,15 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     }
 
     // Stage 1: Convert atomic objects
-    var applicationIdMap = new Dictionary<string, List<string>>(); // used in converting blocks in stage 2
+    var applicationIdMap = new Dictionary<string, List<string>>(); // used in converting blocks in stage 2. keeps track of original app id => resulting new app ids post baking
     var count = 0;
     foreach (var (path, obj) in atomicObjects)
     {
       onOperationProgressed?.Invoke("Converting objects", (double)++count / atomicObjects.Count);
       try
       {
-        var fullLayerName = string.Join(Layer.PathSeparator, path);
-        var layerIndex = cache.TryGetValue(fullLayerName, out int value)
-          ? value
-          : GetAndCreateLayerFromPath(path, baseLayerName, cache);
-
+        var layerIndex = _layerManager.GetAndCreateLayerFromPath(path, baseLayerName);
         var result = _converter.Convert(obj);
-
         var conversionIds = HandleConversionResult(result, obj, layerIndex).ToList();
         foreach (var r in conversionIds)
         {
@@ -155,89 +144,17 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     }
 
     // Stage 2: Convert instances
-    // TODO: do not forget to add to report things
-    var sortedInstanceComponents = instanceComponents
-      .OrderByDescending(x => x.obj.MaxDepth) // Sort by max depth, so we start baking from the deepest element first
-      .ThenBy(x => x.obj is InstanceDefinitionProxy ? 0 : 1) // Ensure we bake the deepest definition first, then any instances that depend on it
-      .ToList();
-    var definitionIdAndApplicationIdMap = new Dictionary<string, int>();
+    var (createdInstanceIds, consumedObjectIds, instanceConversionResults) = _instanceObjectsManager.BakeInstances(
+      instanceComponents,
+      applicationIdMap,
+      baseLayerName,
+      onOperationProgressed
+    );
 
-    count = 0;
-    foreach (var (path, instanceOrDefinition) in sortedInstanceComponents)
-    {
-      onOperationProgressed?.Invoke("Converting blocks", (double)++count / sortedInstanceComponents.Count);
-      try
-      {
-        if (instanceOrDefinition is InstanceDefinitionProxy definitionProxy)
-        {
-          var currentApplicationObjectsIds = definitionProxy.Objects
-            .Select(x => applicationIdMap.TryGetValue(x, out List<string> value) ? value : null)
-            .Where(x => x is not null)
-            .SelectMany(id => id)
-            .ToList();
-
-          var definitionGeometryList = new List<GeometryBase>();
-          var attributes = new List<ObjectAttributes>();
-
-          foreach (var id in currentApplicationObjectsIds)
-          {
-            var docObject = doc.Objects.FindId(new Guid(id));
-            definitionGeometryList.Add(docObject.Geometry);
-            attributes.Add(docObject.Attributes);
-          }
-
-          // POC: Currently we're relying on the definition name for identification if it's coming from speckle and from which model; could we do something else?
-          var defName = $"{baseLayerName} ({definitionProxy.applicationId})";
-          var defIndex = doc.InstanceDefinitions.Add(
-            defName,
-            "No description", // POC: perhaps bring it along from source? We'd need to look at ACAD first
-            Point3d.Origin,
-            definitionGeometryList,
-            attributes
-          );
-
-          // POC: check on defIndex -1, means we haven't created anything - this is most likely an recoverable error at this stage
-          if (defIndex == -1)
-          {
-            throw new ConversionException("Failed to create an instance defintion object.");
-          }
-
-          if (definitionProxy.applicationId != null)
-          {
-            definitionIdAndApplicationIdMap[definitionProxy.applicationId] = defIndex;
-          }
-
-          // Rhino deletes original objects on block creation - we should do the same.
-          doc.Objects.Delete(currentApplicationObjectsIds.Select(stringId => new Guid(stringId)), false);
-          bakedObjectIds.RemoveAll(id => currentApplicationObjectsIds.Contains(id));
-          conversionResults.RemoveAll(
-            conversionResult =>
-              conversionResult.ResultId != null && currentApplicationObjectsIds.Contains(conversionResult.ResultId) // note: as in rhino created objects are deleted, highlighting them won't work
-          );
-        }
-
-        if (
-          instanceOrDefinition is InstanceProxy instanceProxy
-          && definitionIdAndApplicationIdMap.TryGetValue(instanceProxy.DefinitionId, out int index)
-        )
-        {
-          var transform = MatrixToTransform(instanceProxy.Transform, instanceProxy.Units);
-          var layerIndex = GetAndCreateLayerFromPath(path, baseLayerName, cache);
-          var id = doc.Objects.AddInstanceObject(index, transform, new ObjectAttributes() { LayerIndex = layerIndex });
-          if (instanceProxy.applicationId != null)
-          {
-            applicationIdMap[instanceProxy.applicationId] = new List<string>() { id.ToString() };
-          }
-
-          bakedObjectIds.Add(id.ToString());
-          conversionResults.Add(new(Status.SUCCESS, instanceProxy, id.ToString(), "Instance (Block)"));
-        }
-      }
-      catch (Exception ex) when (!ex.IsFatal())
-      {
-        conversionResults.Add(new(Status.ERROR, instanceOrDefinition as Base ?? new Base(), null, null, ex));
-      }
-    }
+    bakedObjectIds.RemoveAll(id => consumedObjectIds.Contains(id)); // remove all objects that have been "consumed"
+    bakedObjectIds.AddRange(createdInstanceIds); // add instance ids
+    conversionResults.RemoveAll(result => result.ResultId != null && consumedObjectIds.Contains(result.ResultId)); // remove all conversion results for atomic objects that have been consumed (POC: not that cool, but prevents problems on object highlighting)
+    conversionResults.AddRange(instanceConversionResults); // add instance conversion results to our list
 
     // Stage 3: Return
     return new(bakedObjectIds, conversionResults);
@@ -310,41 +227,6 @@ public class RhinoHostObjectBuilder : IHostObjectBuilder
     var groupIndex = _contextStack.Current.Document.Groups.Add(groupName, objectIds);
     var group = _contextStack.Current.Document.Groups.FindIndex(groupIndex);
     return group;
-  }
-
-  // POC: This is the original DUI3 function, this will grow over time as we add more conversions that are missing, so it should be refactored out into an ILayerManager or some sort of service.
-  private int GetAndCreateLayerFromPath(string[] path, string baseLayerName, Dictionary<string, int> cache)
-  {
-    var currentLayerName = baseLayerName;
-    RhinoDoc currentDocument = _contextStack.Current.Document;
-
-    var previousLayer = currentDocument.Layers.FindName(currentLayerName);
-    foreach (var layerName in path)
-    {
-      currentLayerName = baseLayerName + Layer.PathSeparator + layerName;
-      currentLayerName = currentLayerName.Replace("{", "").Replace("}", ""); // Rhino specific cleanup for gh (see RemoveInvalidRhinoChars)
-      if (cache.TryGetValue(currentLayerName, out int value))
-      {
-        previousLayer = currentDocument.Layers.FindIndex(value);
-        continue;
-      }
-
-      var cleanNewLayerName = layerName.Replace("{", "").Replace("}", "");
-      var newLayer = new Layer { Name = cleanNewLayerName, ParentLayerId = previousLayer.Id };
-      var index = currentDocument.Layers.Add(newLayer);
-      cache.Add(currentLayerName, index);
-      previousLayer = currentDocument.Layers.FindIndex(index); // note we need to get the correct id out, hence why we're double calling this
-    }
-    return previousLayer.Index;
-  }
-
-  [Pure]
-  private static string[] GetLayerPath(TraversalContext context)
-  {
-    string[] collectionBasedPath = context.GetAscendantOfType<Collection>().Select(c => c.name).ToArray();
-    string[] reverseOrderPath =
-      collectionBasedPath.Length != 0 ? collectionBasedPath : context.GetPropertyPath().ToArray();
-    return reverseOrderPath.Reverse().ToArray();
   }
 
   // POC: Not too proud of this being here, will be moving soon to the instance manager
